// models/booking.model.ts
export interface Booking {
    id: string;
    user_id: string;
    user_name: string;
    user_lastname: string;
    room_id: string;
    room_name: string;
    title: string;
    description: string;
    phone: string;
    start_time: number;
    end_time: number;
    status: string;
    approved_by: string;
    nameapproved_by: string;
    created_at: number;
    updated_at: number;
    }